Extract ConnectivityBreakAnalysis from DcSensitivityAnalysis
